dts: bcm2711: EMMC2 can address the whole first GB
commit  b655917e6adbdfe39aff32db9d028280fdea69f4 from
https://github.com/raspberrypi/linux.git rpi-5.4.y

Although 0xfc000000 looks like an inaccessible RAM address (due to the
peripheral mappings), with RAM mapped at 0xc0000000 (as it is on the
30/32-bit VPU bus) this is actually 0x3c000000 in the ARM memory space,
which is fine.

This difference is potentially the cause of some warnings seen in
sdhci_send_command.

Fixes: "dts: bcm2711: Move emmc2 to its own 'bus'"

Signed-off-by: Phil Elwell <phil@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>