MLK-23852 net: wireless: nxp: mxm_wifiex: add ignore "-Wpacked-bitfield-compat"
commit  from
https://source.codeaurora.org/external/imx/linux-imx

The mxm_wifiex driver define bit-fields of type "char" in
many strunctures, GCC informs you when the offset of such
a field has changed in GCC 4.4 that may lead to differences
in the structure layout. But it doesn't impact the driver,
so use -Wno-packed-bitfield-compat to disable this warning.

Reviewed-by: tian yang <yang.tian@nxp.com>
Signed-off-by: Fugang Duan <fugang.duan@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>