staging: vc04_services: isp: Make all references to bcm2835_isp_fmt const
commit  6a138ef2603aa82fa06efdd87f8f6be94adf2f08 from
https://github.com/raspberrypi/linux.git rpi-5.6.y

The array of potential formats and their configuration should be const.
Rework all accesses so that this is possible.

The list of supported formats was taking a copy of entries from this table.
This is unnecessary, therefore allocate an array of pointers instead of
an array of entries.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>