staging: vc04_services: isp: Remove duplicated initialisation
commit  5b51c8335805e88739c46600af5999da6a9c9269 from
https://github.com/raspberrypi/linux.git rpi-5.8.y

With the codec code from which this was derived, the driver had to
get the supported formats for both input and output ports.
This had been copied across, however here we have independent nodes
for each port, but the code had been left in to do the same thing
twice.
Remove the duplicate.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>