drm/vc4: Fix VIC usage with Broadcast RGB
commit  7e4d963bb28eecec699ccda00ecaf495f6780d5c from
https://github.com/raspberrypi/linux.git rpi-5.4.y

Adding the Broadcast RGB range selection broke the VIC
field of the AVI infoframes on HDMI, zeroing them for all
modes on an HDMI monitor.

Correct this so that it is only zeroed if the range is
contrary to the standard range of the mode.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>