staging:vc04_services: bcm2835-codec: Request headers with I-frame
commit  efd513f91df32b18a2fa68cd9d2a33d8e6545f41 from
https://github.com/raspberrypi/linux.git rpi-5.8.y

V4L2 wishes to have the codec header bytes in the same buffer as the
first encoded frame, so it does become 1-in 1-out for encoding.
The firmware now has an option to do this, so enable it.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>