staging:vc04_services: bcm2835-codec: Request headers with I-frame
commit  f6d3e5641a0254ec58072e32fdf5fc6ca98b4590 from
https://github.com/raspberrypi/linux.git rpi-5.6.y

V4L2 wishes to have the codec header bytes in the same buffer as the
first encoded frame, so it does become 1-in 1-out for encoding.
The firmware now has an option to do this, so enable it.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>