staging:vc04_services: bcm2835-camera: Request headers with I-frame
commit  6d70d85e5232f2f60ed447f747aa7c2c52bfc9b7 from
https://github.com/raspberrypi/linux.git rpi-5.8.y

V4L2 wishes to have the codec header bytes in the same buffer as the
first encoded frame, so it does become 1-in 1-out for encoding.
The firmware now has an option to do this, so enable it.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>