staging: vchiq_arm: Clean up 40-bit DMA support
commit  eb32708e238f05cc32314d9c0e1101de762460f4 from
https://github.com/raspberrypi/linux.git rpi-5.4.y

Manage the split between addresses for the VPU and addresses for the
40-bit DMA controller with a dedicated DMA device pointer that on non-
BCM2711 platforms is the same as the main VCHIQ device. This allows
the VCHIQ node to stay in the usual place in the DT, and removes the
ugly VC_SAFE macros.

Signed-off-by: Phil Elwell <phil@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>