media: i2c: imx290: Add support for V4L2_CID_VBLANK
commit  7f864f1572ca300b0654a3442b61424160191331 from
https://github.com/raspberrypi/linux.git rpi-5.4.y

In order to calculate framerate and durations userspace needs
the vertical blanking information. This can be configurable,
and indeed the datasheet lists different values for VBLANK for
the 1080p and 720p modes.

Add the new control, and adopt the datasheet values for each mode.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>