MLK-24355 arm: dts: imx7ulp-evk: fix remoteproc boot
commit  from
https://source.codeaurora.org/external/imx/linux-imx

81fb06cb6bd4 ("remoteproc: Fall back to using parent memory pool if no dedicated available")
breaks i.MX7ULP SD Boot.

i.MX7ULP not use vdevbuffer for communication between Acore and Mcore.
However we still use memory-region for vring. Because we added
"shared-dma-pool" to the resered region, so when vdevbuffer not found,
the memory-region will be used for the virtio data buffer.

So to fix the issue, we could remove "shared-dma-buffer" to avoid
the area to be used to data buffer, then it will still use global
dma area for data buffer.

Tested-by: Haibo Chen <haibo.chen@nxp.com>
Signed-off-by: Peng Fan <peng.fan@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>